+++ display_student_identifiers.py
@@ -1,0 +1,19 @@
+""" GatorGrouper randomly assigns a list of students to groups """
+
+from random import shuffle
+import argparse
+import itertools
+import sys
+
+DEFALT_STUDENT_FILE = "students.txt"
+GATORGROUPER_HOME = "GATORGROUPER_HOME"
+
+DEFAULT_TEAM_SIZE = 2
+SINGLETON_GROUP = 1
+
+
+def display_student_identifiers(student_identifers):
+    """ Display the student identifiers """
+    for student in student_identifers:
+        print(student)
+
