+++ gatorgrouper.py
@@ -4,103 +4,22 @@ from random import shuffle
 import argparse
 import itertools
 import sys
+#function files
+from parse_gatorgrader_arguments import parse_gatorgrader_arguments
+from read_student_file import read_student_file
+from display_diagnostics import display_diagnostics
+from display_student_identifiers import display_student_identifiers
+from display_student_groups import display_student_groups
+from shuffle_students import shuffle_students
+from group_students import group_students
+from display_welcome_message import display_welcome_message
+
 
 DEFALT_STUDENT_FILE = "students.txt"
 GATORGROUPER_HOME = "GATORGROUPER_HOME"
 
 DEFAULT_TEAM_SIZE = 2
 SINGLETON_GROUP = 1
-
-
-def parse_gatorgrader_arguments(args):
-    """ Parses the arguments provided on the command-line """
-    gg_parser = argparse.ArgumentParser(
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-
-    gg_parser.add_argument(
-        "--verbose",
-        help="Display verbose diagnostic information",
-        action="store_true")
-
-    gg_parser.add_argument(
-        "--group-size",
-        help="Number of students in a group",
-        type=int,
-        default=DEFAULT_TEAM_SIZE,
-        required=False)
-
-    gg_parser.add_argument(
-        "--students-file",
-        help="File containing last name of students",
-        type=str,
-        default=DEFALT_STUDENT_FILE,
-        required=False)
-
-    gg_arguments_finished = gg_parser.parse_args(args)
-    return gg_arguments_finished
-
-
-def read_student_file(students_file_name):
-    """ Reads the student identifies from the specific file,
-        returning the identifiers in a list """
-    with open(students_file_name, 'r') as students_file:
-        student_identifers = [line.strip() for line in students_file]
-    return student_identifers
-
-
-def display_diagnostics(student_identifers, student_groups):
-    """ Display information about what was generated """
-    print("Successfully placed",
-          len(student_identifers), "students into",
-          len(student_groups), "groups")
-    print()
-
-
-def display_student_identifiers(student_identifers):
-    """ Display the student identifiers """
-    for student in student_identifers:
-        print(student)
-
-
-def display_student_groups(student_groups):
-    """ Display the student groups with labels """
-    group_counter = 1
-    for student_group in student_groups:
-        print("Group", group_counter)
-        print(*student_group)
-        print()
-        group_counter = group_counter + 1
-
-
-def shuffle_students(student_identifers):
-    """ Shuffle the student identifiers """
-    shuffled_student_identifers = student_identifers[:]
-    shuffle(shuffled_student_identifers)
-    return shuffled_student_identifers
-
-
-def group_students(student_identifers, group_size):
-    """ Group the student identifiers """
-    iterable = iter(student_identifers)
-    # use itertools to chunk the students into groups
-    student_groups = list(
-        iter(lambda: list(itertools.islice(iterable, group_size)), []))
-    # merge a single student into the previous group
-    last_group_index = len(student_groups) - 1
-    if len(student_groups[last_group_index]) == SINGLETON_GROUP:
-        receiving_group = student_groups[last_group_index - 1]
-        too_small_group = student_groups[last_group_index]
-        receiving_group.append(*too_small_group)
-        student_groups.remove(too_small_group)
-    return student_groups
-
-
-def display_welcome_message():
-    """ Display a welcome message """
-    print()
-    print("GatorGrouper: Automatically Assign Students to Groups")
-    print("https://github.com/gkapfham/gatorgrouper")
-    print()
 
 
 if __name__ == '__main__':
@@ -131,3 +50,4 @@ if __name__ == '__main__':
     display_diagnostics(shuffled_student_identifers,
                         grouped_student_identifiers)
     display_student_groups(grouped_student_identifiers)
+
