+++ parse_gatorgrader_arguments.py
@@ -1,0 +1,41 @@
+""" GatorGrouper randomly assigns a list of students to groups """
+
+from random import shuffle
+import argparse
+import itertools
+import sys
+
+DEFALT_STUDENT_FILE = "students.txt"
+GATORGROUPER_HOME = "GATORGROUPER_HOME"
+
+DEFAULT_TEAM_SIZE = 2
+SINGLETON_GROUP = 1
+
+
+def parse_gatorgrader_arguments(args):
+    """ Parses the arguments provided on the command-line """
+    gg_parser = argparse.ArgumentParser(
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+
+    gg_parser.add_argument(
+        "--verbose",
+        help="Display verbose diagnostic information",
+        action="store_true")
+
+    gg_parser.add_argument(
+        "--group-size",
+        help="Number of students in a group",
+        type=int,
+        default=DEFAULT_TEAM_SIZE,
+        required=False)
+
+    gg_parser.add_argument(
+        "--students-file",
+        help="File containing last name of students",
+        type=str,
+        default=DEFALT_STUDENT_FILE,
+        required=False)
+
+    gg_arguments_finished = gg_parser.parse_args(args)
+    return gg_arguments_finished
+
