+++ shuffle_students.py
@@ -1,0 +1,20 @@
+""" GatorGrouper randomly assigns a list of students to groups """
+
+from random import shuffle
+import argparse
+import itertools
+import sys
+
+DEFALT_STUDENT_FILE = "students.txt"
+GATORGROUPER_HOME = "GATORGROUPER_HOME"
+
+DEFAULT_TEAM_SIZE = 2
+SINGLETON_GROUP = 1
+
+
+def shuffle_students(student_identifers):
+    """ Shuffle the student identifiers """
+    shuffled_student_identifers = student_identifers[:]
+    shuffle(shuffled_student_identifers)
+    return shuffled_student_identifers
+
